services/worker-kafka/requirements.txt
kafka-python==2.0.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
